feat: add spec level features, restructure crate modules
