+++ rdm-core/src/parameter_traits/mod.rs
@@ -6,7 +6,6 @@ pub mod core_impl;
 pub mod heapless_impl; // We use heapless for convenient no_std collections with fixed capacity
 #[cfg(feature = "std")]
 pub mod std_impl;
-
 pub trait RdmParameterData: Sized {
     fn size_of(&self) -> usize;
