+++ src/rdm/parameter/mod.rs
@@ -1,5 +1,9 @@
 pub mod e120;
+#[cfg(feature = "e133")]
 pub mod e133;
+#[cfg(feature = "e137_1")]
 pub mod e137_1;
+#[cfg(feature = "e137_2")]
 pub mod e137_2;
+#[cfg(feature = "e137_7")]
 pub mod e137_7;
